Add the capability to trace all the functions.
`(clj-libssh2.logging/init)` will cause all of the functions in the
library to log their inputs and outputs at level :trace. This can
produce a *lot* of output, so it's advisable to configure logging
carefully before enabling this.

Both `clj-libssh2.logging/trace-fn` and `clj-libssh2.logging/trace-ns`
can be used for more fine grained tracing of this library.